Refactor user services and enhance error handling
Deleted UserService.js and UserServiceFindAll.js in the frontend and incorporated their functionalities into UserxService.java, with the addition of the saveUserx method. Optimized fetch request in App.js to implement error handling for server accessibility issues. Expanded and updated unit tests in UserxControllerTest.java to reflect the new changes and dependencies.